% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_data-documentation.R
\docType{data}
\name{AgeRaceDR}
\alias{AgeRaceDR}
\title{Death rates in the United States by age and race, 2013}
\format{
A \code{data.frame} with 360 observations and 4 variables.
\describe{
 \item{age}{The exact age, in years, at which life expectany is calculated.}
 \item{race}{The racial group of the deceased One of \code{all} all races; \code{white} 
 whites; \code{black} black / African-American; \code{hispanic} Hispanic; \code{asian.isl}
 Asian and Pacific Islander; \code{nat.amer} Native American or Alaska Native.}
 \item{gender}{The gender of the deceased. One of \code{c(both, male, female)}}
 \item{death_rate}{The raw death rate. See details.}
}
}
\source{
\url{http://www.cdc.gov/nchs/nvss/deaths.htm}
}
\usage{
AgeRaceDR
}
\description{
A dataset containing death rates for individuals by age group and race 
for the United States, 2013.
}
\section{Details}{

\itemize{
\item{The death rate is defined as deaths per 100,000 population.} 
}
}

\references{
Xu, J. Q., S. L. Murphy, and K. D. Kochanek. "Deaths: final data for 2013." National 
Vital Statistics Reports 64.2 (2015).
}
\keyword{datasets}
