% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_class-methods_helpers.R
\name{all_geog_constraint_race}
\alias{all_geog_constraint_race}
\title{Create race constraint list to a set of geographies}
\usage{
all_geog_constraint_race(obj)
}
\arguments{
\item{obj}{An object of class \code{"synthACS"}.}
}
\description{
Create a new race constraint list to the mapping between a a set 
of macro datasets and a matching set of micro dataset (supplied as class 'synthACS').
}
\seealso{
\code{\link{all_geogs_add_constraint}}
}

