% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_data-documentation.R
\docType{data}
\name{BR2014}
\alias{BR2014}
\title{Birth Rates by Age and Race of Mother}
\format{
A \code{data.frame} with 1,750 observations and 4 variables.
\describe{
 \item{year}{The year for which data was was recorded.}
 \item{race}{The racial group of the mothers. One of \code{all} all races; \code{white} 
 non-hispanic whites; \code{black_aa} black / African-American; \code{nat_amer} American Indian
 or Native Alaskan; \code{asian_isl} Asian or Pacific Islander; \code{hisp_lat} Hispanic or Latin 
 American.}
 \item{age_group}{The age group of the mother.}
 \item{birth_rate}{The birth rate. See Details.}
}
}
\source{
\url{http://www.cdc.gov/nchs/nvss/births.htm}
}
\usage{
BR2014
}
\description{
A dataset containing birth rate data in the United States by age and race of the 
mother. Data for all races is provided for 1970-2014 and for individual races from 1989-2014.
}
\section{Details}{

\itemize{
\item{The birth rate is defined as births per 1,000 women in the specified group (age and race).} 
\item{Populations are based on census counts enumerated as of April 1 of the census year and estimated
as of July 1 for non-census years.}
\item{Beginning in 1997, birth rates for age group 45up by relating births to all women age 45 or
older to this group. Prior to 1997, only births to women age 45-49 were included.} 
}
}

\references{
Hamilton, Brady E., et al. "Births: final data for 2014." National Vital Statistics 
Reports 64.12 (2015): 1-64.
}
\keyword{datasets}
