% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_class-methods_helpers.R
\name{combine_smsm}
\alias{combine_smsm}
\title{Combine separate SMSM optimizations}
\usage{
combine_smsm(...)
}
\arguments{
\item{...}{A list of objects of class 'smsm_set'.}
}
\description{
Combine objects of class "smsm_set" into a single object of class "smsm_set"
}
\examples{
\dontrun{
 combined <- combine_smsm(smsm1, smsm2, smsm3)
}
}
\seealso{
\code{\link[synthACS]{split}}, \code{\link[synthACS]{all_geog_optimize_microdata}}
}
