% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_class-methods_helpers.R
\name{endyear}
\alias{endyear}
\title{Get the endyear from a "macroACS" object.}
\usage{
endyear(acs)
}
\arguments{
\item{acs}{An object of class \code{"macroACS"}.}
}
\description{
Get the data collection endyear from a "macroACS" object
}

