% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_class-methods_helpers.R
\name{fetch_data}
\alias{fetch_data}
\title{Get Aggregate Data Specified Geography}
\usage{
fetch_data(acs, geography, dataset = c("estimate", "st.err"), choice = NULL)
}
\arguments{
\item{acs}{An object of class \code{"macroACS"}.}

\item{geography}{A character vector allowing string matching via \code{\link[base]{grep}} to 
a set of specified geographies. All values may be specified by \code{"*"}.}

\item{dataset}{Either \code{"estimate"} or \code{"st.err"}. Do you want data on estimated 
population counts or estimated standard errors?}

\item{choice}{A character vector specifying the name of one of the datasets in \code{acs}}
}
\description{
Gets aggregate, macro, data, either estimate or standard error, for a specified geography
and specified dataset.
}
