% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_class-methods_helpers.R
\name{gen_attr_vectors}
\alias{gen_attr_vectors}
\title{Generate attribute vectors}
\usage{
gen_attr_vectors(acs, choice)
}
\arguments{
\item{acs}{An object of class \code{"macroACS"}.}

\item{choice}{A character vector specifying the name of one of the datasets in \code{acs}}
}
\description{
Generate a list of attribute vectors for new synthetic attribute creation from a 
"macroACS" object.
}
\seealso{
\code{\link{all_geog_synthetic_new_attribute}}, \code{\link{synthetic_new_attribute}}
}
