% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_class-methods_helpers.R
\name{get_dataset_names}
\alias{get_dataset_names}
\title{Get dataset names from a "macroACS" object.}
\usage{
get_dataset_names(acs)
}
\arguments{
\item{acs}{An object of class \code{"macroACS"}.}
}
\description{
Get the names of the datasets in a given "macroACS" object.
}
\seealso{
\code{\link{fetch_data}}
}
