% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_class-methods_helpers.R
\name{get_geography}
\alias{get_geography}
\title{Get the geography title from a "macroACS" object.}
\usage{
get_geography(acs)
}
\arguments{
\item{acs}{An object of class \code{"macroACS"}.}
}
\description{
Get the summary information of the geography selected from a "macroACS" object
}
