% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_class-methods_helpers.R
\name{get_ind_income}
\alias{get_ind_income}
\title{Get Individual Income for Specified Geography}
\usage{
get_ind_income(acs, geography, dataset = c("estimate", "st.err"))
}
\arguments{
\item{acs}{An object of class \code{"macroACS"}.}

\item{geography}{A character vector allowing string matching via \code{\link[base]{grep}} to 
a set of specified geographies. All values may be specified by \code{"*"}.}

\item{dataset}{Either \code{"estimate"} or \code{"st.err"}. Do you want data on estimated 
population counts or estimated standard errors?}
}
\description{
Gets data, either estimate or standard error, on a specified geography's population 
income in the past 12 months.
}

