% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_class-methods_helpers.R
\name{get_span}
\alias{get_span}
\title{Get the span from a "macroACS" object.}
\usage{
get_span(acs)
}
\arguments{
\item{acs}{An object of class \code{"macroACS"}.}
}
\description{
Get the data collection span from a "macroACS" object
}
