% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_class-methods_helpers.R
\name{is.macro_micro}
\alias{is.macro_micro}
\title{Check macro_micro class}
\usage{
is.macro_micro(x)
}
\arguments{
\item{x}{any R object.}
}
\value{
Returns \code{TRUE} if its argument has class "macro_micro" among its classes and
\code{FALSE} otherwise.
}
\description{
Function that checks if the target object is a \code{macro_micro} object.
}
