% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_class-methods_helpers.R
\name{is.macroACS}
\alias{is.macroACS}
\title{Check macroACS class}
\usage{
is.macroACS(x)
}
\arguments{
\item{x}{any R object.}
}
\value{
Returns \code{TRUE} if its argument has class "macroACS" among its classes and
\code{FALSE} otherwise.
}
\description{
Function that checks if the target object is a \code{macroACS} object.
}
