% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_class-methods_helpers.R
\name{is.micro_synthetic}
\alias{is.micro_synthetic}
\title{Check micro_synthetic class}
\usage{
is.micro_synthetic(x)
}
\arguments{
\item{x}{any R object.}
}
\value{
Returns \code{TRUE} if its argument has class "micro_synthetic" among its classes and
\code{FALSE} otherwise.
}
\description{
Function that checks if the target object is a \code{micro_synthetic} object.
}
