% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_class-methods_helpers.R
\name{is.synthACS}
\alias{is.synthACS}
\title{Check synthACS class}
\usage{
is.synthACS(x)
}
\arguments{
\item{x}{any R object.}
}
\value{
Returns \code{TRUE} if its argument has class "synthACS" among its classes and
\code{FALSE} otherwise.
}
\description{
Function that checks if the target object is a \code{synthACS} object.
}
