% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_data-documentation.R
\docType{data}
\name{la_hospitals}
\alias{la_hospitals}
\title{Hospitals in Los Angeles County, CA USA}
\format{
A \code{data.frame} with 631 observations and 7 variables
\describe{
  \item{geo_long}{The hospital's longitude.}
  \item{geo_lat}{The hospital's lattitude.}
  \item{city}{The hospital's postal city.}
  \item{state_fips}{The hospital's alpha FIPS code.}
  \item{zip}{The hospital's five digit postal ZIP code.}
  \item{census_tract}{The census tract in which the hospital is located.}
  \item{county_name}{The hospital's county -- "LOS ANGELES".}
}
}
\usage{
la_hospitals
}
\description{
An anonymized dataset containing the geographic information of hospitals in Los 
Angeles County California, USA.
}
\keyword{datasets}
