% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_data-documentation.R
\docType{data}
\name{LifeExp}
\alias{LifeExp}
\title{Life expectancy at certain ages; United States, 2013}
\format{
A \code{data.frame} with 396 observations and 4 variables.
\describe{
 \item{age}{The exact age, in years, at which life expectany is calculated.}
 \item{race}{The racial group of the deceased One of \code{all} all races; \code{white} 
 whites; \code{black} black / African-American; \code{hispanic} Hispanic; \code{non.hisp.white}
 non Hispanic whites; \code{non.hispanic.black} non Hispanic blacks.}
 \item{gender}{The gender of the deceased. One of \code{c(both, male, female)}}
 \item{life_expectancy}{The life expectancy for an individual at the exact age with the given 
 race and gender.}
}
}
\source{
\url{http://www.cdc.gov/nchs/nvss/deaths.htm}
}
\usage{
LifeExp
}
\description{
A dataset containing life expectancy at certain ages by race, hispanic origin and sex 
for the United States, 2013.
}
\references{
Xu, J. Q., S. L. Murphy, and K. D. Kochanek. "Deaths: final data for 2013." National 
Vital Statistics Reports 64.2 (2015).
}
\keyword{datasets}
