% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_data-documentation.R
\docType{data}
\name{rawDR}
\alias{rawDR}
\title{Raw Death Rate by race and gender}
\format{
A \code{data.frame} with 612 observations and 4 variables.
\describe{
 \item{year}{The year for which data was was recorded.}
 \item{race}{The racial group of the deceased One of \code{all} all races; \code{white} 
 whites; \code{black_aa} black / African-American; \code{nat_amer} American Indian
 or Native Alaskan; \code{asian_isl} Asian or Pacific Islander; \code{hisp_lat} Hispanic.}
 \item{gender}{The gender of the deceased. One of \code{c(both, male, female)}}
 \item{death_rate}{The raw death rate. See details.}
}
}
\source{
\url{http://www.cdc.gov/nchs/nvss/deaths.htm}
}
\usage{
rawDR
}
\description{
A dataset containing raw death rate data by race and gender of the deceased. Data  
is provided for 1980-2013.
}
\section{Details}{

\itemize{
\item{The death rate is defined as deaths per 100,000 population.} 
\item{Populations are based on census counts enumerated as of April 1 of the census year and estimated
as of July 1 for non-census years.}
}
}

\references{
Xu, J. Q., S. L. Murphy, and K. D. Kochanek. "Deaths: final data for 2013." National 
Vital Statistics Reports 64.2 (2015).
}
\keyword{datasets}
