% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_class-methods_helpers.R
\name{split}
\alias{split}
\title{Split a "macroACS" object}
\usage{
split(acs, n_splits)
}
\arguments{
\item{acs}{An object of class \code{"macroACS"}.}

\item{n_splits}{An integer for the number of splits you wish to create.}
}
\description{
Split a "macroACS" object into subsets. This may be helpful for users who have
limited memory available on their machines before proceding to derive sample synthetic micro data.
}
\seealso{
\code{\link{derive_synth_datasets}}
}
