% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_class-methods_helpers.R
\name{summary}
\alias{summary}
\title{Summarizing SMSM fits}
\usage{
summary(object)
}
\arguments{
\item{object}{An object of class \code{'smsm_set'}, typically a result of call to 
\code{\link{all_geog_optimize_microdata}}}
}
\description{
\code{summary} method for class 'smsm_set'.
}

