% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_class-methods_helpers.R
\name{summary.smsm_set}
\alias{summary.smsm_set}
\title{Summarizing SMSM fits}
\usage{
\method{summary}{smsm_set}(object, ...)
}
\arguments{
\item{object}{An object of class \code{'smsm_set'}, typically a result of call to 
\code{\link{all_geog_optimize_microdata}}}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
\code{summary} method for class 'smsm_set'.
}
